requirements.txt
requests==2.32.3
PyYAML==6.0.3
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
